% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isocountry}
\alias{isocountry}
\title{Country names with ISO country codes}
\format{
A data frame with 249 rows and 16 variables:
\describe{
\item{name}{country name}
\item{alpha_2}{ISO alpha-2 code}
\item{alpha_3}{ISO alpha-3 code}
\item{country_code}{numeric code}
\item{iso_3166_2}{ISO 3166-2 subdivision code}
\item{is_independent}{is the country independent}
\item{region_code}{UN region code}
\item{region_name}{UN region name}
\item{subregion_code}{UN subregion code}
\item{subregion_name}{UN subregion name}
\item{intermediate_region_code}{UN intermediate region code}
\item{intermediate_region_name}{UN intermediate region name}
\item{m49_code}{UN M49 code}
\item{oecd_member}{is the country an OECD member}
\item{eu_member}{is the country an EU member}
\item{emu_member}{is the country an EMU member}
}
}
\source{
\itemize{
\item \url{https://en.wikipedia.org/wiki/ISO_3166-1}
\item \url{https://unstats.un.org/unsd/methodology/m49/overview}
\item \url{https://en.wikipedia.org/wiki/OECD}
\item \url{https://en.wikipedia.org/wiki/Member_state_of_the_European_Union}
}
}
\usage{
isocountry
}
\description{
ISO 3166-1 country codes with full country names, UN regional classifications,
OECD membership status, EU membership status, and Economic and Monetary Union (EMU) membership status.
}
\examples{
isocountry
}
\keyword{datasets}
