% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isocurrency}
\alias{isocurrency}
\title{ISO currency codes}
\format{
A data frame with 269 rows and 4 variables:
\describe{
\item{country}{country name}
\item{currency}{currency name}
\item{code}{ISO currency code}
\item{number}{ISO currency number}
}
}
\source{
\itemize{
\item \url{https://www.iban.com/currency-codes}
}
}
\usage{
isocurrency
}
\description{
ISO 4217 currency codes with full country names.
}
\keyword{datasets}
