% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isocurrency}
\alias{isocurrency}
\title{ISO currency codes}
\format{
A data frame with 261 rows and 6 variables:
\describe{
\item{currency_name}{ISO currency name}
\item{currency_code}{ISO currency code}
\item{currency_number}{ISO currency number}
\item{country_name}{ISO country name}
\item{country_code}{ISO alpha-2 country code}
\item{country_number}{ISO country number}
}
}
\source{
\url{https://www.iban.com/currency-codes}
}
\usage{
isocurrency
}
\description{
ISO 4217 currency codes with full country names.
}
\examples{
isocurrency
}
\keyword{datasets}
